% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtra_d.R
\docType{data}
\name{arqi}
\alias{arqi}
\title{Filtra pela distancia entre um timestep e outro. Tambem calcula deslocamento total de cada SCM.}
\format{
An object of class \code{character} of length 1.
}
\usage{
arqi
}
\arguments{
\item{arqi}{Arquivo de entrada: c_YYYY.csv}

\item{arqo}{Arquivo de saida: d_YYYY.csv}

\item{dist_max}{Maxima distancia permitida em km entre um passo de tempo e outro.
No mestrado usei 300 km a cada 30 min.}

\item{coords}{Nomes das colunas das coordenas longitude e latitude (nesta ordem)}
}
\value{
data.frame
}
\description{
A funcao filtra_d
}
\examples{
\dontrun{
filtra_d(arqi = "../c_200101.csv")
}
}
\keyword{datasets}
