% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_density.R
\name{calc_density}
\alias{calc_density}
\title{Calculate density maps}
\usage{
calc_density(ifile, nx, ny, h, lims, islist = FALSE)
}
\arguments{
\item{ifile}{Character. Input filename. Generally as c_YYYYMM.txt}

\item{nx}{Integer. Bandwidth for x direction.}

\item{ny}{Integer. Bandwidth for x direction.}

\item{h}{Numeric. This is the bandwith. By default it is 1.5}

\item{lims}{Vector. The limits of the rectangle covered by the grid.}

\item{islist}{Logical. If FALSE (default), returns a RasterBrick.
If TRUE, returns a list of RasterBricks.}
}
\value{
RasterBrick. If islist = TRUE, return list of RasterBricks
}
\description{
Calculate density maps along all the lifecycle by YEAR and MONTH
}
\examples{
\dontrun{
calc_density(ifile = "d_2000_2010.csv",
             nx = 800,                         # Rehbein (2021) 
             ny = 800,                         # Rehbein (2021)
             h = 2,                            # Rehbein (2021)
             lims = c(c(-82, -40), c(-23, 7))) # para a bacia Amazonica
}
}
