% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtra_b.R
\name{filtra_b}
\alias{filtra_b}
\title{Filter MCSs by area of interest}
\usage{
filtra_b(
  ifile,
  ofile,
  aoi,
  focean = TRUE,
  continental_shp = system.file("extdata", "SA/sa.shp", package = "percolator")
)
}
\arguments{
\item{ifile}{Character. Input filename. Ex: a_YYYY.csv}

\item{ofile}{Character. Output filename. Ex: b_YYYY.csv}

\item{aoi}{Character. Set the name of the shapefile of the 
region to filter the MCSs. Amazon basin by default.}

\item{focean}{Logical. Should I indicate the oceanic MCSs? TRUE by default.}

\item{continental_shp}{Character. Set the name of the shapefile for filtering 
oceanic MCSs. South America contours by default.}
}
\value{
data.table
}
\description{
This function reads the output from filtra_a function, filter MCSs
over an area of interest (for instance Amazon basin), classifies the MCSs by
their origin (for instance oceanic or continental), and then returns the output 
in data.table format or saved according to the ofile argument.
}
\examples{
\dontrun{
# Do not run
filtra_b(ifile = "/media/amanda/Elements/AR/SAIDAS_FORTRACC/fam/IMERG/a_200101010000.txt", 
         ofile = "/media/amanda/Elements/AR/SAIDAS_FORTRACC/fam/IMERG/b_200101010000.txt",
         aoi = system.file("extdata",  
                           "BACIA/bacia_amazonica_contorno.shp",
                           package = "percolator"), 
         focean = FALSE)
}
}
