% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtra_c.R
\name{filtra_c}
\alias{filtra_c}
\title{Filtro pela curva de crescimento, onde a maxima extensao eh o pico da curva.
Esta funcao tambem identifica as fases (genesis, maturation, dissipation) dos SCMs..}
\usage{
filtra_c(arqi, arqo)
}
\arguments{
\item{arqi}{Arquivo de entrada: b_YYYY.csv}

\item{arqo}{Arquivo de saida: c_YYYY.csv}
}
\value{
data.frame
}
\description{
Esta funcao le os arquivos filtrados pela bacia e filtra pela curva de crescimento.
}
\examples{
\dontrun{
filtra_c(arqi = "../b_200101.csv")
}
}
