% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtra_extensions.R
\name{filtra_extent}
\alias{filtra_extent}
\title{Filter by minimum extention (precipitation Houze)}
\usage{
filtra_extent(ifile, ofile, pixels_size)
}
\arguments{
\item{ifile}{Character. Input filename. Generally as c_YYYYMM.txt}

\item{ofile}{Character. Output filename. Sugested: d_YYYYMM.txt}

\item{pixels_size}{Numeric. Pixels size in km2. Change from date to date.}
}
\value{
data.frame
}
\description{
Houze --> MCSs with area greater than 100 km in at least one direction.
To be used only with precipitation data.
}
\examples{
\dontrun{
filtra_extent(ifile = "/media/amanda/Elements/AR/SAIDAS_FORTRACC/fam/IMERG/d_200101010000.csv", 
         ofile = "/media/amanda/Elements/AR/SAIDAS_FORTRACC/fam/IMERG/APAGAR.txt",
         pixels_size = 11*11)
}
}
