% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtra_saag.R
\name{filtra_saag}
\alias{filtra_saag}
\title{Filter Fortracc outputs to match MCSs definition from the SAAG Deep Convection Group}
\usage{
filtra_saag(
  xmin = -81.95,
  xmax = -34.05,
  ymin = -55.95,
  ymax = 12.95,
  ncols = 480,
  nlins = 690,
  type = "Observational",
  family_file,
  pathi_to_prec_file = "/glade/work/arehbein/SAAG/3yr/IMERG/",
 
    pathi_to_fortracc_clusters = "/glade/work/arehbein/SAAG/3yr/MCSs/OBS/SAIDAS_FORTRACC_Tb/clusters/",
 
    pathi_to_fortracc_txt = "/glade/work/arehbein/SAAG/3yr/MCSs/OBS/SAIDAS_FORTRACC_Tb/diag.txt/",
  pathi_to_masks_files = "/glade/work/arehbein/SAAG/3yr/MCSs/OBS/MCSs_MASKs/"
)
}
\arguments{
\item{xmin}{Numeric. Western coord. of the native data. EX:
ncdump merg_2010010100_4km-pixel.nc | less}

\item{xmax}{Numeric. Eastern coord. of the native data. EX:}

\item{ymin}{Numeric. Southern coord. of the native data. EX:}

\item{ymax}{Numeric. Northern coord. of the native data. EX:}

\item{ncols}{Integer. Relative to the native data.}

\item{nlins}{Integer. Relative to the native data.}

\item{type}{Character. Use "Observational" for GPM IMERG or "WRF" for WRF SAAG Simulations.}

\item{family_file}{Character. Name of the family file (including the path to it)
obtained by Fortracc. Must be one starting with "fam_"}
}
\value{
netCDF
}
\description{
This function filter the ForTraCC outputs to select MCSs that match 
the criteria defined by the SAAG Deep Convection Working Group at April 13, 2022 meeting 
(https://ral.ucar.edu/projects/south-america-affinity-group-saag/deep-convection-working-group).  
It also creates unique ID's for each family, obtain their lifespan, timeUTC, timeLT, 
phases, distances, and total displacement. There is an option for filter 
(if growing_curve is TRUE) data eliminating families in which the
genesis, maturation, and/or dissipation happened at the same time. 
The criterias are: 
- Minimum Tb <= 241 K area must be at least 40,000 km2 for at least 4 continuous hours;
- Minimum peak precipitation - 10 mm/h at least one pixel (0.1 x 0.1 deg for 1 h)
  inside the whole MCS for 4 continuous hours;
- 20,000 km mm/h (e.g., 100 km x 100 km x 2 mm/h) minimum rainfall volume at 
  least once in the lifetime of the MCS;
- Overshoot threshold - Tb < 225 K exists during the cloud lifetime.
}
\examples{
\dontrun{
filtra_saag(xmin = -81.95, # Coords. pegas do ncdump --> centroide da celula de grade
            xmax = -34.05,
            ymin = -55.95,
            ymax = 12.95,
            ncols = 480,
            nlins = 690,
            type = "Observational",
            family_file = "/glade/work/arehbein/SAAG/3yr/MCSs/OBS/SAIDAS_FORTRACC_Tb/diag.txt/fam_SAAG_Tb_GPM_IMERG_1106_s2.txt",
            pathi_to_prec_file = "/glade/work/arehbein/SAAG/3yr/IMERG/",
            pathi_to_fortracc_clusters = "/glade/work/arehbein/SAAG/3yr/MCSs/OBS/SAIDAS_FORTRACC_Tb/clusters/",
            pathi_to_fortracc_txt = "/glade/work/arehbein/SAAG/3yr/MCSs/OBS/SAIDAS_FORTRACC_Tb/diag.txt/",
            pathi_to_masks_files = "/glade/work/arehbein/SAAG/3yr/MCSs/OBS/MCSs_MASKs/")
}
}
