% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mcs_csv.R
\name{get_mcs_csv}
\alias{get_mcs_csv}
\title{Post-processing ForTraCC outputs to obtain MCSs using the SAAG Deep Convection Group criteria}
\usage{
get_mcs_csv(family_file, ofile_csv)
}
\arguments{
\item{family_file}{Character. Path to the family file obtained by Fortracc. Must start with "fam_".}

\item{ofile_csv}{Character. Output CSV file path for filtered data.}
}
\value{
csv
}
\description{
The function performs the following steps:

- Filtering by Tb (brightness temperature) thresholds.
- Filtering by size criteria.
- Filtering by precipitation (PCP) criteria.
}
\details{
This function identifies and filters Mesoscale Convective Systems (MCSs) based on specific criteria 
established by the SAAG Deep Convection Working Group. It creates final plain text dataset
for analysis and visualization.
}
\examples{
\dontrun{
get_mcs_masks(family_file = "fam_SAAG_Tb_pcp_info_intermedio_FINAL_2001-2001_01.csv",
              ofile_csv = "MCSs_MASK/MCS_SAAG_WRF_p_2000.csv"
) }
}
