% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mcs_masks.R
\name{get_mcs_masks}
\alias{get_mcs_masks}
\title{Post-processing ForTraCC outputs to obtain MCSs using the SAAG Deep Convection Group criteria}
\usage{
get_mcs_masks(
  year_start,
  month_start,
  day_start,
  hour_start,
  year_end,
  month_end,
  day_end,
  hour_end,
  ncols,
  nlins,
  family_file,
  path_to_fortracc_clusters,
  cluster_prefix,
  ofile_mask,
  attribute = "Tb MCSs Masks from ForTraCC-percolator",
  reference_file = system.file("extdata/merg_2011010100_4km-pixel.nc", package =
    "percolator")
)
}
\arguments{
\item{year_start}{Integer. Starting year for filtering the data.}

\item{month_start}{Integer. Starting month for filtering the data.}

\item{day_start}{Integer. Starting day for filtering the data.}

\item{hour_start}{Integer. Starting hour for filtering the data.}

\item{year_end}{Integer. Ending year for filtering the data.}

\item{month_end}{Integer. Ending month for filtering the data.}

\item{day_end}{Integer. Ending day for filtering the data.}

\item{hour_end}{Integer. Ending hour for filtering the data.}

\item{ncols}{Integer. Number of columns relative to the native data.}

\item{nlins}{Integer. Number of lines relative to the native data.}

\item{family_file}{Character. Path to the family file obtained by get_mcs_csv.R}

\item{path_to_fortracc_clusters}{Character. Path to the directory containing Fortracc clusters.}

\item{cluster_prefix}{Character. Prefix for cluster files.}

\item{ofile_mask}{Character. Output NetCDF file path for masks.}

\item{attribute}{Character. Description of the attribute for the NetCDF file.}

\item{reference_file}{Character. Path to the reference file.}
}
\value{
netCDF
}
\description{
The function performs the following steps:

- Filtering by Tb (brightness temperature) thresholds.
- Filtering by size criteria.
- Filtering by precipitation (PCP) criteria.
- Data processing and preparation.
- Mask creation for MCSs.
}
\details{
This function identifies and filters Mesoscale Convective Systems (MCSs) based on specific criteria 
established by the SAAG Deep Convection Working Group. It creates masks for further analysis or visualization.
}
\examples{
\dontrun{
get_mcs_masks(year_start = 2001,
            month_start = 1,
            day_start = 1,
            hour_start = 0,
            year_end = 2001,
            month_end = 1,
            day_end = 31,
            hour_end = 23,
            ncols = 480,
            nlins = 690,
            family_file = "MCSs_MASK/MCS_SAAG_WRF_p_2000.csv",
            path_to_fortracc_clusters = "clusters/",
            cluster_prefix = "WRF_20y_p_",
            ofile_mask = "MCSs_MASK/2000_p_WRF_SAAG-MCS-mask-file.nc",
            attribute = "Description of the attribute",
            reference_file = system.file("extdata/merg_2011010100_4km-pixel.nc", 
                                         package = "percolator")
) }
}
