% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prec_under_tb.R,
%   R/get_prec_under_tb_with_MERGES.R
\name{get_prec_under_tb}
\alias{get_prec_under_tb}
\title{Calculate Precipitation related to the ForTraCC Tb objects}
\usage{
get_prec_under_tb(
  month,
  year_start,
  year_end,
  ncols,
  nlins,
  type,
  cluster_prefix,
  family_file,
  path_to_prec_file,
  path_to_fortracc_clusters,
  path_to_masks_files,
  area_file = system.file("extdata/grid_area_saag/area_wrf_km2.nc", package =
    "percolator")
)

get_prec_under_tb(
  month,
  year_start,
  year_end,
  ncols,
  nlins,
  type,
  cluster_prefix,
  family_file,
  path_to_prec_file,
  path_to_fortracc_clusters,
  path_to_masks_files,
  area_file = system.file("extdata/grid_area_saag/area_wrf_km2.nc", package =
    "percolator")
)
}
\arguments{
\item{month}{Integer. Month of the tracking.}

\item{year_start}{Integer. Starting year of the data.}

\item{year_end}{Integer. Ending year of the data.}

\item{ncols}{Integer. Number of columns in the native data.}

\item{nlins}{Integer. Number of lines in the native data.}

\item{type}{Character. Type of data used for tracking the MCSs. It will be 
added at the dataset for identification purposes and will be used to select the
corresponding precipitation field. Available options: "IMERGE" for GPM IMERG or 
"WRF" for WRF SAAG Simulations.}

\item{cluster_prefix}{Character. Prefix for image files.}

\item{family_file}{Character. Path to the family file.}

\item{path_to_prec_file}{Character. Path to the precipitation files.}

\item{path_to_fortracc_clusters}{Character. Path to the ForTraCC cluster files.}

\item{path_to_masks_files}{Character. Path to the masks files.}

\item{area_file}{Character. Path to the area file.}
}
\value{
A csv file.

A csv file.
}
\description{
This function calculates the precipitation area and volume under
the cloud shield. It overlap the brightness temperature (Tb) objects (clusters) 
generated by ForTracc with the corresponding precipitation field.

This function calculates the precipitation area and volume under
the cloud shield. It overlap the brightness temperature (Tb) objects (clusters) 
generated by ForTracc with the corresponding precipitation field.
}
\examples{
\dontrun{
get_prec_under_tb(month = 1,
            ncols = 480,
            nlins = 690,
            type = "WRF",
            year_start = 2000,
            year_end = 2000,
            cluster_prefix = "WRF_20y_p_",
            area_file = system.file("extdata/grid_area_saag/area_wrf_km2.nc", package = "percolator"),
            family_file = "diag.txt/fam_SAAG_WRF_20yr_p_0112_s2.txt",
            path_to_prec_file = paste0("DATA/WRF/PREC_ACC_NC/"),
            path_to_fortracc_clusters = "clusters/",
            path_to_masks_files = "MCSs_MASKs/")
} 
\dontrun{
get_prec_under_tb(month = 1,
            ncols = 480,
            nlins = 690,
            type = "WRF",
            year_start = 2000,
            year_end = 2000,
            cluster_prefix = "WRF_20y_p_",
            area_file = system.file("extdata/grid_area_saag/area_wrf_km2.nc", package = "percolator"),
            family_file = "diag.txt/fam_SAAG_WRF_20yr_p_0112_s2.txt",
            path_to_prec_file = paste0("DATA/WRF/PREC_ACC_NC/"),
            path_to_fortracc_clusters = "clusters/",
            path_to_masks_files = "MCSs_MASKs/")
} 
}
