% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prec_under_tb_mm.R
\name{get_prec_under_tb_mm}
\alias{get_prec_under_tb_mm}
\title{Calculate Precipitation related to the ForTraCC Tb objects}
\usage{
get_prec_under_tb_mm(
  month,
  year_start,
  year_end,
  ncols,
  nlins,
  type,
  family_file,
  cluster_prefix,
  pcp_file,
  pcp_varname,
  pcp_time_name = "time",
  path_to_fortracc_clusters,
  path_to_masks_files,
  plot = FALSE,
  area_file = system.file("extdata/grid_area_saag/area_wrf_km2.nc", package =
    "percolator")
)
}
\arguments{
\item{month}{Integer. Month of the tracking.}

\item{year_start}{Integer. Starting year of the data.}

\item{year_end}{Integer. Ending year of the data.}

\item{ncols}{Integer. Number of columns in the native data.}

\item{nlins}{Integer. Number of lines in the native data.}

\item{type}{Character. Type of data used for tracking the MCSs. It will be 
added at the dataset for identification purposes.}

\item{family_file}{Character. Path to the family file.}

\item{cluster_prefix}{Character.}

\item{pcp_file}{Character. Precipitation monthly file name. 
Ex: "pathi_to_prec/200101_PREC_ACC_NC-on_IMERG_grid.nc"}

\item{pcp_varname}{Character. Precipitation variable name. Ex: "PREC_ACC_NC"}

\item{pcp_time_name}{Character. Time variable name. Default: "time"}

\item{path_to_fortracc_clusters}{Character. Path to the ForTraCC cluster files.}

\item{path_to_masks_files}{Character. Path to the masks files.}

\item{plot}{Logical. If clusters over precipitation should be plotted at the
path_to_masks_files. See figs directory inside that path. Default: FALSE.}

\item{area_file}{Character. Path to the area file.}
}
\value{
A csv file.
}
\description{
This function calculates the precipitation area and volume under
the cloud shield. It overlap the brightness temperature (Tb) objects (clusters) 
generated by ForTracc with the corresponding precipitation field. Requires monthly
input files with hourly precipitation.
}
\examples{
\dontrun{
get_prec_under_tb(month = 1,
            ncols = 480,
            nlins = 690,
            type = "WRF",
            year_start = 2000,
            year_end = 2000,
            cluster_prefix = "WRF_20y_p_",
            area_file = system.file("extdata/grid_area_saag/area_wrf_km2.nc", package = "percolator"),
            family_file = "diag.txt/fam_SAAG_WRF_20yr_p_0112_s2.txt",
            path_to_prec_file = paste0("DATA/WRF/PREC_ACC_NC/"),
            path_to_fortracc_clusters = "clusters/",
            path_to_masks_files = "MCSs_MASKs/")
} 
}
