% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_LT.R
\name{LT}
\alias{LT}
\title{Calculate the Local Time for each timestep of the MCSs}
\usage{
LT(
  x,
  y = system.file("extdata", "ne_10m_time_zones/ne_10m_time_zones.shp", package =
    "percolator"),
  coords = c("XLON", "YLAT")
)
}
\arguments{
\item{x}{data.table or data.frame with lat and long coordinates.}

\item{y}{String. Input file with the timezones in shapefile format. By default 
y is Natural Earth timezones}

\item{coords}{vector of strings. Coordinates names (longitude first).
\url{www.naturalearthdata.com/downloads/10m-cultural-vectors/timezones/}}
}
\value{
data.frame
}
\description{
Calculate the Local Time for each timestep of the MCSs, 
using (by default) Natural Earth timezones.
}
\examples{
\dontrun{
}
}
