% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percolator.R
\docType{package}
\name{percolator}
\alias{percolator}
\title{percolator: input clouds, output numbers}
\description{
This package organizes functions for filtering ForTraCC outputs, 
obtaining comon features for studying MCSs.
}
\section{Authors}{

- Amanda Rehbein (ORCID: https://orcid.org/0000-0002-8714-7931 -
 mantainer: amanda.rehbein@usp.br)
- Sergio Ibarra Espinosa (ORCID: https://orcid.org/0000-0002-3162-1905)
}

\references{
- Rehbein 
- Sakamoto
- Vila et al
}
