% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_filter.R
\name{pre_filter}
\alias{pre_filter}
\title{Performs a pre-filter for}
\usage{
pre_filter(ifile, ofile, fonte, period)
}
\arguments{
\item{ifile}{Character. Input filename. Generally as YYYYMM.txt}

\item{ofile}{Character. Output filename. Sugested: a_YYYYMM.csv}

\item{fonte}{Character. Source of the data. For instance: "Tb" for the
MERG product, "IMERG", "CMORPH", "NICAM AMIP", "NICAM HighResMIP".
It will be added as a field in the final data.frame. It is not 
used for filtering.}

\item{period}{Character. Choose between "present" or "future". 
It will be added as a field in the final data.frame. 
It is not used for filtering.}
}
\value{
data.table
}
\description{
This function performs a pre-filter after adding essentials fields
to the data. It reads the files ouputed by fortran routine/function
(YYYYMM.txt); give the fonte and period for each family; creates IDs for each family,
obtain their lifespan, timeUTC, phases, distances, and total displacement; 
filter data eliminating families in which the genesis, maturation, and/or dissipation 
happened at the same time; and then save the final dataset.
}
\examples{
\dontrun{
pre_filter(ifile = "/media/amanda/Elements/AR/SAIDAS_FORTRACC/fam/IMERG/200101010000.txt", 
           ofile = "/media/amanda/Elements/AR/SAIDAS_FORTRACC/fam/IMERG/p_200101010000.txt",
           fonte = "IMERG",
           period = "present")
}
}
