% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tb.R
\name{read_tb}
\alias{read_tb}
\title{Transform ForTraCC outputs from block format to plain text}
\usage{
read_tb(ifile, ofile)
}
\arguments{
\item{ifile}{character. Name of input file. Generally starts
with "fam". Maximum of 100 characters.}

\item{ofile}{character. Name of output file. For running the other filters,
it must be "csv" format. Maximum of 100 characters.}
}
\value{
character
}
\description{
This function transform the ForTraCC output 
(from tb) that has a block format (with head 
and tail information) in a plain text, with no blank lines
in between the families. It takes  information from the
header and write it as many time as the duration (or TIME).

This function calls a fortran subroutine for reading ForTraCC
"fam" archives outuput and rewrite them in .csv format (or other).
This routine must not be used with files with more than 40000 families!!!
}
\examples{
\dontrun{
read_tb(ifile = "/home/amanda/Documents/filtra_scms/TESTE/tb_fam1403_s2.txt",
        ofile = "/home/amanda/Documents/filtra_scms/TESTE/tb_out.txt")
}
}
