% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_UTC.R
\name{UTC}
\alias{UTC}
\title{Lê arquivos output do read_family.R e filtra pela bacia,
separando por oceano e continente}
\usage{
UTC(ANO, MES, DIA, HORAINI, TIME)
}
\arguments{
\item{ANO}{ano}

\item{MES}{mes}

\item{DIA}{dia}

\item{HORAINI}{hora inicial do SCM em UTC}

\item{TIME}{horas passadas apos o inicio do sistema. Ex: 0, 0.5, 1.0, 1.5}
}
\value{
data.frame
}
\description{
This function reads the output from read_family.R to filter MCSs
Over the AMazon basin
}
\examples{
\dontrun{
# Do not run Isso vai aparecer na documentação
head(df)
hist(df$DURACAO) #
}
}
