% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_nc.R
\name{write_nc}
\alias{write_nc}
\title{Save the MASK files from the filtra_saag function}
\usage{
write_nc(type, dt, pathi_to_prec_file, pathi_to_masks_files)
}
\arguments{
\item{type}{Character. Use "Observational" for GPM IMERG or "WRF" for WRF SAAG Simulations.}

\item{dt}{data.table containing the information of date and years}
}
\value{
netCDF
}
\description{
Save the Tb MCSs mask files in netCDF format.
}
\examples{
\dontrun{
write_nc(type = "Observational",
         dt = dt,
         pathi_to_prec_file = "/glade/work/arehbein/SAAG/3yr/IMERG/",
         pathi_to_masks_files = "/glade/work/arehbein/SAAG/3yr/MCSs/OBS/MCSs_MASKs/")
}
}
