% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtra_a.R
\name{filtra_a}
\alias{filtra_a}
\title{Lê arqvuivos fam_*.csv, coloca o header e retorna arq a_YYYY.csv para
ser filtrado pela bacia, curva e deslocamento.}
\usage{
filtra_a(
  arqi,
  arqo,
  fonte,
  variavel,
  epoca,
  header = FALSE,
  filtra_time = 3,
  classifica_lifespan = 5.7
)
}
\arguments{
\item{arqi}{Arquivo de entrada. Se for "OLR" tera formato fam_YYYYMM.csv;
Se for "PCP" fam_R12534354_YYYYMMDDHHMM.csv)}

\item{arqo}{Arquivo de saida: a_YYYY.csv}

\item{fonte}{Fonte de dados: "CMORPH", "NICAM AMIP" ou "NICAM HigResMIP"}

\item{variavel}{Qual variavel dos dados de entrada para o ForTraCC. 
Escolher entre "PCP" (precipitacao) e "OLR"}

\item{epoca}{Escolher entre "presente" ou "futuro"}

\item{header}{Boolean. Default is FALSE. Has the arqi file a header?}

\item{filtra_time}{SCMs devem ter no minimo filtra_time horas de duracao.
Em geral, SCMs sao maiores ou iguais a 3 horas}

\item{classifica_lifespan}{Limiar para classificar SCMs em short- ou long-lived}
}
\value{
data.frame
}
\description{
This function reads the fam_*.csv files,
excludes the white spaces between the families,
gives the names to the columns, creates IDs, and
save the new data.frame.
}
\examples{
\dontrun{
filtra_a(arqi = "200001.csv", 
         arqo = "a_200001.csv",
         fonte = "CMORPH",
         variavel = "PCP",
         epoca = "presente",
         filtra_time = 3)
}
}
