% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtra_b.R
\name{filtra_b}
\alias{filtra_b}
\title{Filtra SCMs que interagiram em alguma regiao de interesse. Por exemplo: bacia Amazonica.
Tambem separa oceanicos e continentais.}
\usage{
filtra_b(
  arqi,
  arqo,
  focean = TRUE,
  pregion = system.file("extdata", "BACIA/bacia_amazonica_contorno.shp", package =
    "fortracc.deal"),
  psa = system.file("extdata", "SA/sa.shp", package = "fortracc.deal")
)
}
\arguments{
\item{arqi}{Arquivo de entrada: a_YYYY.csv}

\item{arqo}{Arquivo de saida: b_YYYY.csv}

\item{focean}{Logical. Should I filter ocean MCSs?}

\item{pregion}{Character. Indica o nome do shapefile da regiao a filtrar os scms}

\item{psa}{Character. Indica o nome do shapefile da America do Sul}
}
\value{
data.frame
}
\description{
This function reads the output from read_family.R to filter MCSs
Over the AMazon basin
}
\examples{
\dontrun{
# Do not run Isso vai aparecer na documentação
filtra_b(arqi = "../a_200101.csv",
         arqo = "../b_200101.csv")
}
}
