% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtra_d.R
\name{filtra_d}
\alias{filtra_d}
\title{Calculate and filter MCSs displacement}
\usage{
filtra_d(ifile, ofile, dist_max = 300, coords = c("XLON", "YLAT"))
}
\arguments{
\item{ifile}{Arquivo de entrada: c_YYYY.csv}

\item{ofile}{Arquivo de saida: d_YYYY.csv}

\item{dist_max}{Maxima distancia permitida em km entre um passo de tempo e outro.
No mestrado usei 300 km a cada 30 min.}

\item{coords}{Nomes das colunas das coordenas longitude e latitude (nesta ordem)}
}
\value{
data.frame
}
\description{
Filtra pela distancia entre um timestep e outro. Tambem
 calcula deslocamento total de cada SCM.
}
\examples{
\dontrun{
filtra_d(ifile = "/media/amanda/Elements/AR/SAIDAS_FORTRACC/fam/IMERG/c_200101010000.txt", 
         ofile = "/media/amanda/Elements/AR/SAIDAS_FORTRACC/fam/IMERG/d_200101010000.txt")
}
}
