% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtra_saag.R
\name{filtra_saag}
\alias{filtra_saag}
\title{Obtain SAAG Deep Convection Working Group MCSs from ForTraCC outputs}
\usage{
filtra_saag(
  ifile,
  ofile,
  dx = 0.1 * 111.32,
  dy = 0.1 * 111.32,
  min_volume = 20000,
  variable = "PCP",
  experiment = "WRF SAAG 3yr",
  period = "PRESENT",
  growing_curve = FALSE
)
}
\arguments{
\item{ifile}{Character. Input filename. Generally as fam_WRF_PCP_WRF_YYYYMMDDHHMM.txt}

\item{ofile}{Character. Output filename. Sugested: WRF_PCP_WRF_YYYY.csv}

\item{dx}{Numeric. Used to determine the area of the pixel. E.g.: 11.132 (default)}

\item{dy}{Numeric. Used to determine the area of the pixel. E.g.: 11.132 (default)}

\item{min_volume}{Numeric. Used for filter as determined in the SAAG DCWG Meeting 
at April 13, 2022.}

\item{variable}{Character. "PCP" (default) or "Tb".}

\item{experiment}{Character. Sugested: "WRF SAAG 3yr" (default) or "WRF SAAG 20yr"}

\item{period}{Character. Choose between "PRESENT" (default) or "FUTURE".}

\item{growing_curve}{Boolean. If TRUE (default) families in which the
genesis, maturation, and/or dissipation happened at the same time are removed.
It is a filter.}
}
\value{
data.table
}
\description{
This function filter the ForTraCC outputs to select MCSs that match 
the criterias defined by the SAAG Deep Convection Working Group at April 13, 2022 meeting 
(https://ral.ucar.edu/projects/south-america-affinity-group-saag/deep-convection-working-group).  
It also creates unique ID's for each family, obtain their lifespan, timeUTC, timeLT, 
phases, distances, and total displacement. There is an option for filter 
(if growing_curve is TRUE) data eliminating families in which the
genesis, maturation, and/or dissipation happened at the same time. 
The criterias are: 
- Minimum Tb <= 241 K area must be at least 40,000 km2 for at least 4 continuous hours;
- Minimum peak precipitation - 10 mm/h at least one pixel (0.1 x 0.1 deg for 1 h)
  inside the whole MCS for 4 continuous hours;
- 20,000 km mm/h (e.g., 100 km x 100 km x 2 mm/h) minimum rainfall volume at 
  least once in the lifetime of the MCS;
- Overshoot threshold - Tb < 225 K exists during the cloud lifetime.
}
\examples{
\dontrun{
filtra_saag(ifile <- "/home/amanda/Documents/pos_doc/FAPESP/SAAG/diag.txt/fam_SAAG_PCP_WRF_201006052300.txt", 
           ofile = "/home/amanda/Documents/pos_doc/FAPESP/SAAG/diag.txt/SAAG_PCP_WRF_2010.txt")
}
}
