% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pcp.R
\name{read_pcp}
\alias{read_pcp}
\title{Transform ForTraCC outputs from block format to plain text}
\usage{
read_pcp(filei, fileo)
}
\arguments{
\item{filei}{character. Name of input file. Generally starts
with "fam". Maximum of 100 characters.}

\item{fileo}{character. Name of output file. For running the other filters,
it must be "csv" format. Maximum of 100 characters.}
}
\value{
character
}
\description{
This function transform the ForTraCC output 
(from precipitation) that has a block format (with head 
and tail information) in a plain text, with no blank lines
in between the families. It takes  information from the
header and write it as many time as the duration (or TIME).

This function calls a fortran subroutine for reading ForTraCC
"fam" archives outuput and rewrite them in .csv format (or other).
}
\examples{
\dontrun{
out(filei = "fam_200012.txt",
    fileo = "200012.csv")
}
}
