% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_UTC.R
\name{UTC}
\alias{UTC}
\title{Calculate the time in UTC for each timestep of the MCSs}
\usage{
UTC(YEAR, MONTH, DAY, HOUR, TIME)
}
\arguments{
\item{YEAR}{String. Year field name.}

\item{MONTH}{String. Month field name.}

\item{DAY}{String. Day field name.}

\item{HOUR}{String. Hour of genesis field name.}

\item{TIME}{String. TIME counter lifespan field. E.g.: 0, 0.5, 1.0, 1.5.}
}
\value{
data.frame
}
\description{
Calculate the time in UTC for each timestep of the MCSs
}
\examples{
\dontrun{
dt$timeUTC <- UTC(YEAR = dt$YEAR,
                  MONTH = dt$MONTH,
                  DAY = dt$DAY,
                  HOUR = dt$HOUR,
                  TIME = dt$TIME)
}
}
