% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_nc.R
\name{write_nc}
\alias{write_nc}
\title{Save Network Common Data Form (NetCDF) files}
\usage{
write_nc(
  attribute,
  values,
  date_ini = "2001-01-01 00:00:00",
  ofile = "ofile.nc",
  ntime,
  reference_file = system.file("extdata", "merg_2011010100_4km-pixel.nc", package =
    "percolator"),
  latname = "lat",
  lonname = "lon",
  unit = "",
  longname = "Tb MCSs Masks",
  variable_name = "mcs_mask",
  author_name = "Amanda Rehbein",
  institution_name = "Climate Group of Studies (GrEC)/University of Sao Paulo (USP)",
  additional_comments =
    "Machado et al. (1998), Vila et al. (2008), Rehbein et al. (2020)"
)
}
\arguments{
\item{attribute}{String. Any information to serve as an attribute. Ex:
"Observational Tb (GPM MERGEIR) MCSs Masks from ForTraCC-percolator"}

\item{values}{Vector with numeric or integer values to be written in the
netCDF file.}

\item{date_ini}{String. First date for the TIME attribute. EX: "2001-01-01 00:00:00"}

\item{ofile}{Character. Character with the name of the output file.}

\item{ntime}{Integer. Number of timesteps for the TIME dimension. 
EX: ntime = 365*24.}

\item{precision}{is the prec in ?ncvar_get.}
}
\value{
netCDF
}
\description{
Export data with TIME, XLAT, XLONG dimensions to a netCDF file.
}
\examples{
\dontrun{

}
}
